Add babel-loader dev dependency to enable multiple source files
In order to have multiple source files, i.e. more than just
index.jsx, need to have babel-loader installed. This is only required
so the component in the master branch can be used with the example app
in the gh-pages branch (babel-loader is not required to transpile multiple
files from the src directory into the lib directory).

Also, babel-loader's peer dependency requirement of webpack is met by
the webpack installed with the example app in the gh-pages branch,
so no need to install webpack with the component.

Note that babel-core is included with babel-cli,
so no need to install it separately.